src/requirements.txt
PyPDF2==3.0.1
selenium==4.8.2
requests==2.28.2
regex==2022.10.31
mysql-connector-python==8.0.32
webdrivermanager==0.10.0
pytest==7.3.0